refactor: dao 테스트 후 롤백 적용
