+++ src/main/java/chess/web/controller/dto/RoomRequestDto.java
@@ -2,12 +2,14 @@ package chess.web.controller.dto;
 
 public class RoomRequestDto {
 
-    public static class TitleAndPassword{
+    public static class TitleAndPassword {
         private String title;
         private String password;
 
         private TitleAndPassword() {
-        };
+        }
+
+        ;
 
         public TitleAndPassword(String title, String password) {
             this.title = title;
@@ -23,10 +25,12 @@ public class RoomRequestDto {
         }
     }
 
-    public static class Password{
+    public static class Password {
         private String password;
 
-        private Password(){}
+        private Password() {
+        }
+
         public Password(String password) {
             this.password = password;
         }
